+++ lib/data/storage/import/import_investment.dart
@@ -4,6 +4,7 @@ import 'package:money/core/widgets/dialog/dialog_button.dart';
 import 'package:money/core/widgets/gaps.dart';
 import 'package:money/data/models/money_objects/investments/investment_types.dart';
 import 'package:money/data/models/money_objects/investments/investments.dart';
+import 'package:money/data/models/money_objects/investments/picker_investment_trade_type.dart';
 import 'package:money/data/models/money_objects/payees/payee.dart';
 import 'package:money/data/models/money_objects/securities/security.dart';
 import 'package:money/data/models/money_objects/transactions/transaction.dart';
@@ -73,7 +74,7 @@ List<Widget> getActionButtons(
           units: inputData.units,
           unitPrice: inputData.amountPerUnit,
           investmentType: inputData.investmentType.index,
-          tradeType: InvestmentTradeType.none.index,
+          tradeType: fromInvestmentType(inputData.investmentType).index,
         );
 
         Data().investments.appendNewMoneyObject(investmentToBeAdded, fireNotification: false);